closes #134
